backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pyjwt[crypto]==2.8.0
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-multipart==0.0.9
svix==1.16.0
supabase==2.6.0
loguru==0.7.2

